+++ test/ponder/value.cpp
@@ -146,7 +146,7 @@ TEST_CASE("Ponder has variant values")
     ponder::Value cstringValue = "1";
     ponder::Value stringValue  = ponder::String("1");
     ponder::Value enumValue    = One;
-    ponder::Value objectValue  = object1;
+    ponder::Value objectValue  = ponder::UserObject::makeRef(&object1);
     
     SECTION("values have type")
     {
@@ -396,7 +396,7 @@ TEST_CASE("Ponder has variant values")
         REQUIRE_THROWS_AS(objectValue.to<ponder::String>(),       ponder::BadType);
         REQUIRE_THROWS_AS(objectValue.to<MyEnum>(),            ponder::BadType);
         REQUIRE(objectValue.to<MyClass>() == object1);
-        IS_TRUE(objectValue.to<ponder::UserObject>() == ponder::UserObject(object1));
+        IS_TRUE(objectValue.to<ponder::UserObject>() == ponder::UserObject(&object1));
 
         REQUIRE(objectValue.isCompatible<bool>() ==           true);
         REQUIRE(objectValue.isCompatible<char>() ==           false);
@@ -409,7 +409,7 @@ TEST_CASE("Ponder has variant values")
         REQUIRE(objectValue.isCompatible<unsigned long>() ==  false);
         REQUIRE(objectValue.isCompatible<float>() ==          false);
         REQUIRE(objectValue.isCompatible<double>() ==         false);
-        REQUIRE(objectValue.isCompatible<ponder::String>() ==    false);
+        REQUIRE(objectValue.isCompatible<ponder::String>() == false);
         REQUIRE(objectValue.isCompatible<MyEnum>() ==         false);
         REQUIRE(objectValue.isCompatible<MyClass>() ==        true);
     }
@@ -499,14 +499,16 @@ TEST_CASE("Ponder has variant values")
         REQUIRE(ponder::Value(1.)      == ponder::Value(1.));
         REQUIRE(ponder::Value("1")     == ponder::Value("1"));
         REQUIRE(ponder::Value(One)     == ponder::Value(One));
-        IS_TRUE(ponder::Value(object1) == ponder::Value(object1));
+        IS_TRUE(ponder::Value(&object1) == ponder::Value(&object1));    // refs are ==
+        IS_TRUE(ponder::Value(object1) != ponder::Value(&object1));     // copies are !=
 
         REQUIRE(ponder::Value(true)    != ponder::Value(false));
         REQUIRE(ponder::Value(1)       != ponder::Value(2));
         REQUIRE(ponder::Value(1.)      != ponder::Value(2.));
         REQUIRE(ponder::Value("1")     != ponder::Value("2"));
         REQUIRE(ponder::Value(One)     != ponder::Value(Two));
-        IS_TRUE(ponder::Value(object1) != ponder::Value(object2));
+        IS_TRUE(ponder::Value(&object1) != ponder::Value(&object2));    // ref
+        IS_TRUE(ponder::Value(object1) != ponder::Value(object2));      // copy
     }
 
     SECTION("values can be compared using less than")
@@ -575,8 +577,8 @@ TEST_CASE("Ponder has variant values")
         REQUIRE(equivalent(ponder::Value(1.),      ponder::Value(1.)) ==      true);
         REQUIRE(equivalent(ponder::Value("1"),     ponder::Value("1")) ==     true);
         REQUIRE(equivalent(ponder::Value(One),     ponder::Value(One)) ==     true);
-        REQUIRE(equivalent(ponder::Value(object1), ponder::Value(object1)) == true);
-        REQUIRE(equivalent(ponder::Value(object1), ponder::Value(object2)) == false);
+        REQUIRE(equivalent(ponder::Value(&object1), ponder::Value(&object1)) == true);
+        REQUIRE(equivalent(ponder::Value(&object1), ponder::Value(&object2)) == false);
 
         REQUIRE((ponder::Value(false) < ponder::Value(true)) == true);
         REQUIRE((ponder::Value(1)     < ponder::Value(2)) ==    true);